Initial version of the surface grid handling module.
The module acts as a wrapper of a Dune grid (currently the FoamGrid grid class is used). The example program 'traversal' illustrates how to iterate over the elements. The wrapper is not complete yet, e.g. most of the functionality of the Geometry class (providing mappings between logical and physical coordinates) is missing.

Example files and tests used to verify the correct operation of CMake have been moved to the (hopefully dead) branch "first-steps".