GridFactory class added.
* GridFactory provides means to create structured Bempp grids and import
grids from Gmsh files.

* The class Dune::StructuredGridFactory has been replaced by a modified
implementation Dune::BemppStructuredGridFactory (might be moved to the
Bempp namespace later).

* The traversal examples have been updated to use Bempp::GridFactory.

* The definition of the numerical type used to represent coordinates
(double) has been moved to the (new) lib/grid/common.hpp file.

Note 1: usage of Bempp::GridFactory is not identical to that of
Dune::GridFactory (the latter is a lower-level class); hence it might be
better to rename Bempp::GridFactory to e.g. Bempp::GridCreator in order
to avoid confusion.

Note 2: A bug has been found in Dune::FoamGrid::FoamGridFactory. In the
declaration of insertBoundarySegment() BoundarySegment<2> should be
changed to BoundarySegment<2, 3>.