Significantly reworked the discrete linear operator classes.
If BEM++ is compiled with Trilinos, they are now derived from
Thyra::LinearOpBase, and thus can be easily used with
Thyra-compatible Trilinos modules. An example is the Stratimikos
module, which provides Thyra-compatible wrappers of the iterative
solvers from the Belos module.

A preconditioner based on the approximate LU decomposition of H
matrices is also available in the AcaApproximateLuInverse class.

The example "examples/trilinos.cpp" demonstrates the solution of
a linear system generated with BEM++ using the GMRES solver from
Belos.

If Trilinos is not available, the discrete linear operator classes
still provide fallback "apply" functions implementing the

y := alpha L x + beta y

operarion (with L being the linear operator, x and y vectors, and
alpha and beta scalar), so the user has the possibility of using
the operators with iterative linear solvers of his/her
choice. Export of dense matrix representations of the
operators (as Armadillo matrices) is also supported.