Removed the function assembleOperator() from LinearOperator and its subclasses.
This function was originally intended to be used for collocation-mode
assembly. This has not been implemented yet, and won't be for the
nearest future.

Historical note: assembleOperator() was return a "vector-valued"
discrete operator, i.e. e.g. for an integral operator with tensorial kernel K it would return a 3D array with elements

K(i,j,k) = \int [K(x^j, y) \dot f_k(y)]_i dy,

where the notation []_i is the ith component of the vector in
brackets, x^j denotes the jth test point and f_k the kth trial
function. However, for most practical purposes -- like the
solution of an integral equation -- this 3D array would anyway
need to be converted to a 2D matrix (by combining the first two
dimensions into one). Thus, the 3D representation is probably not
worth the bother of having to deal with separate hierarchies of
Scalar- and VectorValued linear operators. For this reason the
VectorValued operator classes have now been removed.