Cleaned up the example subdirectory. It contains now only two well-commented examples demonstrating the use of the library for the solution of a Neumann and a Dirichlet problem.
All meshes have been moved to the example/meshes subdirectory. Compile-time mesh selection by means of an enumeration variable has been removed. Instead, the user can specify the mesh file on the command line.

All "temporary" code should now be placed in the "scratch" subdirectory (which by default does not exist, but if it exists and contains a CMakeLists.txt file, it is included in the build sequence). The contents of "scratch" should not normally be uploaded to GitHub (at least not to the master branch).