Refactored the GridFunction class. Added mass-matrix management functionality to the Space class.
Grid functions now store both their expansion coefficients in the basis of
the chosen function space and their scalar products with those basis
functions ("projections"). As a result, multiplication of a LinearOperator
with a GridFunction now works correctly (i.e. the result of the discrete
operation -- multiplication of the weak form by a vector of coefficients --
is treated correctly as a vector of projections, not coefficients, of
a function, and these projections are now used to generate the right
GridFunction).

The Space class stores its mass matrix and inverse mass matrix, and provides
member functions apply(nverse)MassMatrix(argument, result). These are
used by the GridFunction.