Refactored the interface of LinearOperator. Made provisions for optimised assembly of operators with symmetric weak forms.
I hope the interface is now clearer for the user, while maintaining its
flexibility (the user can decide to work primarily with formal linear
operators or with discrete linear operators, whatever suits him or her best).
In each case, the ownership of the assembled weak form is explicitly
indicated in the documentation.