Added the possibility of constructing more complex "elementary" integral operators with integrands of the form
\sum_{i=1}^N w_i a_i(x) K(x, y) b_i(y)

(note: the kernel K is shared between all the terms). This will be useful
for implementing the Helmholtz hypersingular operator and operators
of linear elasticity.

It might actually be a good idea to distinguish between "elementary"
integral operators -- with N = 1 -- and "compound" integral operators --
with general N. Maybe in future...