Introduced the distinction between trial space and the range of the operator.
The constructor of LinearOperator takes now references to three spaces:
domain (formerly testSpace), range (new) and the dual to range (formerly
trialSpace). The constructor of GridFunction takes a reference to the dual
space in addition to one to the primal space.

Note: GridFunction doesn't handle the case of primal_space != dual_space
correctly yet.