Great renaming, part 1.
Renamed LinearOperator and its descendants to BoundaryOperator.
Renamed Potential to PotentialOperator.

As a result, Laplace3dSingleLayerPotentialOperator represents now, indeed,
a potential operator -- an operator that applied to a surface density
function generates a field in the whole space except possibly the surface
on which this density is defined. In contrast, a
Laplace3dSingleLayerBoundaryOperator can be evaluated on the surface;
in particular, it provides a means to calculate its weak form.