Removed a mutex from BoundaryOperator::weakForm().
It seems to me that it is acceptable for this function not to be fully
thread-safe.  There is little need for sharing BoundaryOperator
objects between different threads, since they are lightweight and can
be copied cheaply. As far as I understand, different threads can still
safely share the DiscreteBoundaryOperator objects stored internally in
BoundaryOperators by means of Boost shared_ptrs.