Modified the MKL installer yet again -- reverted to linking with individual MKL libraries (mkl_core, mkl_intel_thread, ...) instead of the single mkl_rt.
It turns out that older versions of MKL, including the one distributed
with Enthought, have a bug in mkl_rt, which affects runtime loading of
libiomp5. This leads to error messages about undefined symbols such as
omp_get_num_procs at runtime.

The recent patch setting LINK_INTERFACE_LIBRARIES to "" is kept. It is
indeed necessary for correct propagation of C++ exceptions to Python.