Uniformised handling of the symmetry and label of abstract operators.
The parameter type has been changed from Symmetry to int (since it can
be a bitwise combination of elements of the Symmetry enumeration type).

For all operators it is now possible to set their symmetry explicitly
via a constructor parameter. Some classes are able to determine the
appropriate symmetry automatically (but it can be overridden).