Modified the installer, changing the location where BEM++ and its dependencies are built. Also, moved installer files from "contrib" to "installer".
Previously, BEM++ dependencies were built in the "contrib" subdirectory of the
source tree (polluting it), while BEM++ itself was built in a directory
specified in the "Bempp/build_dir" option in the setup configuration file.

Now, BEM++ dependencies are built in the "contrib" subdirectory
of the directory specified in the "Main/build_dir" option in the setup
configuration file, while BEM++ itself is built in the "bempp" subdirectory
of the same directory.

This makes it possible to test the installer without overwriting the dependency
build directory one uses for everyday work.