GridFunctions no longer need dualSpace during its construction.
The dualSpace needs to be provided only when it is really needed, i.e.
when calling projections() or setProjections(). The vector of projections
is not stored inside the GridFunction object any more.

For compatibility, the existing interface is preserved, but deprecated.
The only change is that projections() now returns an arma::Col object,
not a reference to it. In most cases, this should not be a problem.