Added a benchmark for vectorized modified Helmholtz SLP operator with a modified data layout.
As opposed to the current version of BEM++, the modified data
structures store the data so that values corresponding to successive
test points are adjacent to each other. However, different components
of vectors (e.g. the x, y and z coordinates of points in 3D space)
are not stored next to each other.