Dune is told not to use BLAS or LAPACK.
Reason: automatic search for BLAS/LAPACK libs by the Dune installer
sometimes leads to problems ("random" libraries detected) and currently
there is no robust way of passing BLAS or LAPACK library paths
to all Dune modules -- we would need to modify Dune configure scripts
to pass these paths to the linker invocation building a test file
depending on libdunecommon. This cannot be done until the next major
release of Dune. So for now it's better to just disable BLAS and LAPACK
in Dune, since it appears only one function in Dune -- eigenValues() --
depends on these libraries, and this function isn't called anywhere
in the Dune modules that we use.