Great renaming: basis -> shapeset.
In particular, the classes Basis and CollectionOfBasisTransformations have
been renamed to Shapeset and CollectionOfShapesetTransformations.

The term *shapeset* denotes a family of *shape functions* defined on a
reference element. Thus, from now on, shape functions live on reference
elements, while basis functions live on physical elements. This usage is
consisted with e.g. Szabo and Babuska's and Solin's.