Added to Space functions explicitly responsible for interpolation.
(We now distinguish between getGlobalDofPositions(), used in clustering,
and getGlobalDofInterpolationPoints(), used in interpolation. Note that
a space might not be interpolatory, in which case
getGlobalDofInterpolationPoints() will throw an exception, whereas
getGlobalDofPositions() can still work, perhaps returning the element centre
for all the basis functions located on an element.

The mechanism is sufficiently general to allow interpolation of vector
(e.g. surface-tangential) functions, although it is not yet implemented
for the particular case of Raviart-Thomas spaces.