Non-backwards-compatible change: separated local assemblers for integral and local operators. In addition, removed the ElementaryAbstractBoundaryOperator class.
Rationale: LocalAssemblerForOperators was an artifical creature: half of
its functions were used only in the assembly of integral operators,
the other half only in the assembly of local operators. This was because
originally we did joint assembly of integral and local operators in ACA
mode and so a common class encompassing both functionalities was needed.
We now know this common assembly is a bad idea, so there is no reason
any more for this design wart.