Adding patches for dune (Issue #58)
dune-localfunction_*: These patches cast 1.0 to the relevant type

Patches are now applied via a cmake script.
The script mimics sed: it applies a regex replace.
The advantage is that the patch should be more robust with respect to changes in the file than using
diffs, as in current implementation. The disadvantage is that it will fail silently if the
implementation changes those patterns. That could be good (bug was corrrected) or bad.
Another advantage is that the scripts is nominally more portable, in the unlikely event that BEM++
is installed on a system without the patch executable.