Fixing RPaths after install
Trying a different approach that should be more robust.
The rpaths are fixed at the end of the install process.
This is done by adding a post-install script.
Because CMake can be awkward, a post-install script is an install script executed in the last
included subdirectory. This subdirectory is created explicitly.