Updating to dune version 2.3.1
External project now compiled via Dune's own cmake.
Also, added umbrella target Dune that depends on all the dune sub-components.
This may be one of the problems affecting parallel builds (#115).

This commit goes partway to #114 (dune update) while still retaining static
linkage (#116). The code compiles, but there is a bug in one of unit test over
the foam grid. Might be that the dimensionality of the problem is not passed
correctly to foamgrid somewhere.