Adds wrappers for spaces
Spaces are defined in python/mako/space.py using a dictionary that specifies
the header, class name, implementation, docstring, and characteristics.
In practice, a fair number of these quantities are guessed from the name of the
class.

As far as the user is concerned, creating a space comes down to:

>>> from bempp import Grid, scalar_spaces
>>> grid = Grid(....)
>>> quadratic = scalar_spaces(grid, 'float32', order=2)

A number of different scalar spaces are available. There are yet no wrappers
for vectorial spaces.