Adding first sketch at boundary operator wrappers (#132)
Wrapping the BoundaryOperator class. The memory where instances of the class
are constructed is owned by C++. It is the largest possible size of all
possible `BoundaryOperator<basis_type, result_type>` instances.
The operators are then constructed in-place according to the basis and result
type in the __init__ method.
Member method wrappers should first look to result_type and basis_type to
figure out how to cast the memory to right instance.