added generic FEM-BEM coupling
should work at least for arbitrary order lagrange elements. Maybe even for different kinds of elements.
the mapping uses eval() of a given FEniCS Function in order to evaluate the function on the needed
quadrature points. As long as the quadrature order is high enough this gives the mapping of FEniCS dofs
to the BEM++ dofs.